requirements.txt
alembic==0.8.7
appdirs==1.4.3
click==6.7
Flask==0.12
Flask-Cors==3.0.2
Flask-Migrate==2.0.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
funcsigs==1.0.2
GeoAlchemy2==0.3.0
itsdangerous==0.24
Jinja2==2.9.5
lorem-ipsum-generator==0.3
Mako==1.0.4
MarkupSafe==1.0
mock==2.0.0
packaging==16.8
pbr==1.10.0
Pillow==3.3.1
psycopg2==2.6.2
pyparsing==2.2.0
python-editor==1.0.1
Shapely==1.5.17
six==1.10.0
SQLAlchemy==1.0.14
Unidecode==0.4.20
Werkzeug==0.12
